+++ PriceAggregator.API/Models/AggregatedPriceModel.cs
@@ -4,5 +4,5 @@ public class AggregatedPriceModel
 {
     public DateTime Time { get; set; }
     
-    public double AggregatedPrice { get; set; }
+    public decimal AggregatedPrice { get; set; }
 }
