+++ PriceAggregator.API/Services/IPriceService.cs
@@ -4,7 +4,7 @@ namespace PriceAggregator.API.Services;
 
 public interface IPriceService
 {
-    Task<double> GetAggregatedPrice(string financialInstrument, DateTime time);
+    Task<decimal> GetAggregatedPrice(string financialInstrument, DateTime time);
     
     Task<List<AggregatedPriceModel>> GetPersistedPrices(string financialInstrument, DateTime start, DateTime end);
 }
